+++ HISTORY.md
@@ -48,6 +48,7 @@
 ### 2022-05-02
 * blocks, threads: programmatically change the type of unused custom blocks
 * blocks, threads: new "scope" choice in block menu dropdown
+* blocks, threads: programmatically change the scope of unused custom blocks
 
 ### 2022-05-01
 * byob: programmatically reduce the number of inputs in a custom block